+++ src/main/java/proguard/classfile/kotlin/KotlinConstants.java
@@ -35,6 +35,10 @@ public class KotlinConstants
 
     public static final char FUNCTION_NAME_MANGLE_SEPARATOR = '-';
 
+    public static final String FUNCTION_NAME_ANONYMOUS = "<anonymous>";
+
+    public static final String METHOD_NAME_LAMBDA_INVOKE = "invoke";
+
     public static final String NAME_KOTLIN_METADATA                   = "kotlin/Metadata";
     public static final String TYPE_KOTLIN_METADATA                   = "Lkotlin/Metadata;";
     public static final String NAME_KOTLIN_ANY                        = "kotlin/Any";