+++ docs/md/releasenotes.md
@@ -3,6 +3,7 @@
 ### Improved
 
 - `ExecutingInvocationUnit` now loads values from static final fields.
+- Initialize Kotlin lambda method references when the JVM method name is `<anonymous>`. (`T16483`)
 
 ### Bug fixes